fix: make sure that "default" is not exported as a resource
