+++ src/resources/contacts.js
@@ -1,4 +1,4 @@
-import createResourceDescription from '.'
+import createResourceDescription from './util/createResourceDescription'
 
 export const contacts = createResourceDescription('contacts', {
   includes: [
