+++ src/resources/customers.js
@@ -1,4 +1,4 @@
-import createResourceDescription from '.'
+import createResourceDescription from './util/createResourceDescription'
 
 export const customers = createResourceDescription('customers', {
   filters: ['external_id', 'match', 'limit'],
