+++ src/resources/employees.js
@@ -1,4 +1,4 @@
-import createResourceDescription from '.'
+import createResourceDescription from './util/createResourceDescription'
 
 export const employees = createResourceDescription('employees', {
   filters: ['email', 'external_id'],
