+++ src/resources/enterpriseSettings.js
@@ -1,4 +1,4 @@
-import createResourceDescription from '.'
+import createResourceDescription from './util/createResourceDescription'
 
 export const enterpriseSettings = createResourceDescription(
   'enterprise_settings'
