+++ src/resources/subjectCategories.js
@@ -1,4 +1,4 @@
-import createResourceDescription from '.'
+import createResourceDescription from './util/createResourceDescription'
 
 export const subjectCategories = createResourceDescription(
   'subject_categories',
